+++ reasoner_validator/biolink/__init__.py
@@ -9,6 +9,7 @@ from pprint import PrettyPrinter
 import logging
 
 from bmt import Toolkit
+from bmt.utils import parse_name
 from linkml_runtime.linkml_model import ClassDefinition, Element
 
 from reasoner_validator.sri.util import is_curie
@@ -437,6 +438,33 @@ class BiolinkValidator(ValidationReporter):
             # TODO: not yet sure what else to do here (if anything...yet)
             attribute_constraints: List = edge['attribute_constraints']
 
+    def validate_qualifier_entry(self, context: str, edge_id: str, qualifiers: List[Dict[str, str]]):
+        for qualifier in qualifiers:
+            qualifier_type_id: str = qualifier['qualifier_type_id']
+            qualifier_value: str = qualifier['qualifier_value']
+            try:
+                if not self.bmt.validate_qualifier(
+                        # TODO: temporary workaround, parse 'qualifier_type_id' to core name
+                        qualifier_type_id=parse_name(qualifier_type_id),
+                        qualifier_value=qualifier_value
+                ):
+                    self.report(
+                        code=f"error.{context}.qualifier.invalid",
+                        edge_id=edge_id,
+                        qualifier_type_id=qualifier_type_id,
+                        qualifier_value=qualifier_value
+                    )
+
+            except Exception as e:
+                # broad spectrum exception to trap anticipated short term issues with BMT validation
+                logger.error(f"BMT validate_qualifier Exception: {str(e)}")
+                self.report(
+                    code=f"error.{context}.qualifier.invalid",
+                    edge_id=edge_id,
+                    qualifier_type_id=qualifier_type_id,
+                    qualifier_value=qualifier_value
+                )
+
     def validate_qualifiers(self, edge_id: str, edge: Dict):
         # Edge qualifiers will only be seen in Biolink 3 data,
         # but with missing 'qualifiers', no validation is attempted
@@ -448,71 +476,11 @@ class BiolinkValidator(ValidationReporter):
             return  # nullable: true... an empty 'qualifiers' array is ok?
         else:
             qualifiers: List = edge['qualifiers']
-            for qualifier in qualifiers:
-                # TODO: how do I validate qualifiers here?
-                #     Qualifier:
-                #       additionalProperties: false
-                #       description: >-
-                #         An additional nuance attached to an assertion
-                #       type: object
-                #       properties:
-                #         qualifier_type_id:
-                #           type: string
-                #           description: >-
-                #             The category of the qualifier, drawn from a hierarchy of qualifier
-                #             slots in the Biolink model (e.g. subject_aspect, subject_direction,
-                #             object_aspect, object_direction, etc).
-                #           example: subject_aspect
-                #           nullable: false
-                #         qualifier_value:
-                #           type: string
-                #           description: >-
-                #             The value associated with the type of the qualifier, drawn from
-                #             a set of controlled values by the type as specified in
-                #             the Biolink model (e.g. 'expression' or 'abundance' for the
-                #             qualifier type 'subject_aspect', etc).
-                #           example: expression
-                #           nullable: false
-                #       required:
-                #         - qualifier_type_id
-                #         - qualifier_value
-                pass
-
-    def is_qualifier(self, name: str) -> bool:
-        """
-        Predicate to test if a Biolink element is Qualifier class element.
-
-        :param name: str LinkML Element identifier
-        :return: True if Qualifier; False otherwise
-        """
-        element: Element = self.bmt.get_element(name)
-        if not element:
-            return False  # TODO: should probably raise an exception here instead?
-        else:
-            # TODO: naive test of existence of qualifier... can we do better?
-            return element.name.endswith("qualifier")
-
-    def is_enum(self, name: str) -> bool:
-        """
-        Predicate to test (by name) if a Biolink element is an enum.
-
-        :param name: str LinkML Element identifier
-        :return: True if enum; False otherwise
-        """
-        element: Element = self.bmt.get_element(name)
-        if not element:
-            return False  # TODO: should probably raise an exception here instead?
-        else:
-            # TODO: naive name-driven test of existence of qualifier...
-            #       Not very robust....can we do better?
-            return element.name.lower().endswith("enum")
-
-    def is_enum_value(self, enum_name: str, value) -> bool:
-        element: Element = self.bmt.get_element(enum_name)
-        if not element:
-            return False  # TODO: should probably raise an exception here instead?
-        else:
-            return False  # ... until implemented?
+            self.validate_qualifier_entry(
+                context="knowledge_graph.edge.qualifiers",
+                edge_id=edge_id,
+                qualifiers=qualifiers
+            )
 
     def validate_qualifier_constraints(self, edge_id: str, edge: Dict):
         # Edge qualifiers will only be seen in Biolink 3 data,
@@ -536,91 +504,11 @@ class BiolinkValidator(ValidationReporter):
                 else:
                     # We have a putative non-empty 'qualifier_set'
                     qualifier_set: List = qualifier_set_entry['qualifier_set']
-                    # we have a putative list of qualifiers?
-                    for qualifier in qualifier_set:
-                        # TODO: This test may not be necessary since
-                        #       TRAPI schema validation should pick it up?
-                        if not (qualifier and isinstance(qualifier, Dict)):
-                            self.report(
-                                code="error.query_graph.edge.qualifier_constraints." +
-                                     "qualifier_set.qualifier.invalid",
-                                edge_id=edge_id
-                            )
-                        else:
-                            qualifier_type_id: str = qualifier['qualifier_type_id']
-                            # TODO: (7 Feb 2023) need to refine & complete implementation of Biolink 3.1.2++
-                            #       qualifier validation (need Biolink Model Toolkit implementation support)
-                            if not qualifier_type_id.startswith("biolink:"):
-                                self.report(
-                                    code="error.query_graph.edge.qualifier_constraints." +\
-                                         "qualifier_set.qualifier.qualifier_type_id.not_biolink_curie",
-                                    edge_id=edge_id,
-                                    identifier=qualifier_type_id
-                                )
-                            # Validate the putative qualifier as *not* being abstract, deprecated or a mixin
-                            biolink_class = self.validate_element_status(
-                                context="query_graph.edge.qualifier",
-                                name=qualifier_type_id
-                            )
-                            if biolink_class:
-                                # First pass here is to check here if the name of the
-                                # 'qualifier_type_id' has string suffix 'qualifier'
-                                if not self.is_qualifier(qualifier_type_id):
-                                    self.report(
-                                        code="error.query_graph.edge.qualifier_constraints." +
-                                             "qualifier_set.qualifier.qualifier_type_id.invalid",
-                                        edge_id=edge_id,
-                                        identifier=qualifier_type_id
-                                    )
-                                else:
-                                    qualifier_value_range: List[str] = \
-                                        self.bmt.get_slot_range(slot_name=qualifier_type_id)
-                                    if not qualifier_value_range:
-                                        self.report(
-                                            code="error.query_graph.edge.qualifier_constraints." +
-                                                 "qualifier_set.qualifier.qualifier_type_id.range.missing",
-                                            edge_id=edge_id,
-                                            identifier=qualifier_type_id
-                                        )
-                                    else:
-                                        # Seems like a legitimate 'qualifier_type_id' with a
-                                        # Biolink defined range so now we attempt to validate
-                                        # the associated qualifier value against that range
-                                        for data_type_name in qualifier_value_range:
-                                            data_type: Optional[Element] = self.bmt.get_element(data_type_name)
-                                            assert data_type, f"Undefined Biolink qualifier range {data_type_name}?"
-                                            qualifier_value: str = qualifier['qualifier_value']
-                                            if self.is_enum(data_type.name):
-                                                # TODO: check here if 'qualifier_value' is known to data_type
-
-                                                if not self.is_enum_value(
-                                                        enum_name=data_type_name,
-                                                        value=qualifier_value
-                                                ):
-                                                    self.report(
-                                                        code="error.query_graph.edge.qualifier_constraints." +
-                                                             "qualifier_set.qualifier.qualifier_value.unknown",
-                                                        edge_id=edge_id,
-                                                        enum_name=qualifier_type_id,
-                                                        value=qualifier_value
-                                                    )
-                                            else:
-                                                # Assume that we have a Biolink categorical data type here
-                                                # and that the qualifier_value is a uriorcurie
-                                                # which maps against the category
-                                                # TODO: data_type is a Biolink Category thus we need to validate
-                                                #       the 'qualifier_value' against the available category mappings, e.g.
-                                                if not self.is_enum_value(
-                                                        enum_name=data_type_name,
-                                                        value=qualifier_value
-                                                ):
-                                                    self.report(
-                                                        code="error.query_graph.edge.qualifier_constraints." +
-                                                             "qualifier_set.qualifier.qualifier_value.unknown",
-                                                        edge_id=edge_id,
-                                                        enum_name=qualifier_type_id,
-                                                        value=qualifier_value
-                                                    )
+                    self.validate_qualifier_entry(
+                        context="query_graph.edge.qualifier_constraints.qualifier_set",
+                        edge_id=edge_id,
+                        qualifiers=qualifier_set
+                    )
 
     def validate_predicate(self, edge_id: str, predicate: str):
         """
@@ -628,7 +516,11 @@ class BiolinkValidator(ValidationReporter):
         :param predicate: putative Biolink Model predicate to be validated
         :return:
         """
-        context: str = f"{self.graph_type.name.lower()}.edge.predicate"
+        graph_type_context: str = self.graph_type.name.lower()
+        if graph_type_context != "input_edge":
+            graph_type_context += ".edge"
+        context: str = f"{graph_type_context}.predicate"
+
         # Validate the putative predicate as *not* being abstract, deprecated or a mixin
         biolink_class = self.validate_element_status(
             context=context,
@@ -638,7 +530,6 @@ class BiolinkValidator(ValidationReporter):
             if not self.bmt.is_predicate(predicate):
                 self.report(
                     code=f"error.{context}.invalid",
-                    context=self.graph_type.value,
                     edge_id=edge_id,
                     predicate=predicate
                 )
@@ -744,7 +635,7 @@ class BiolinkValidator(ValidationReporter):
                     name=category
             )
             if biolink_class and not self.bmt.is_category(category):
-                self.report(code=f"error.{context}.node.category.unknown", category=category)
+                self.report(code=f"error.{context}.node.category.unknown", node_id=node_id, category=category)
                 biolink_class = None
         else:
             self.report(code=f"error.{context}.node.category.missing", node_id=node_id)